CMake: Split sheenbidi into 2 targets
Visual Studio doesn't support a target with sources
that depend on the Debug/Release configuration:

> Target "sheenbidi" has source files which vary by configuration.
> This is not supported by the "Visual Studio 15 2017" generator.

As a workaround, splits adds 2 objects libraries and links one
or the other conditionally.

The libraries are OBJECT rather than STATIC because there is no
point in creating a static archive as they're only ever linked
into the `sheenbidi` library.

Also cleans up the install interface a bit, removing any mentions
of ASAN/UBSAN from the installed config.